tests/requirements-test.txt
# Test dependencies for CAMF test suite

# Core testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0  # For parallel test execution
pytest-timeout>=2.1.0
pytest-mock>=3.11.0
pytest-json-report>=1.5.0

# Test utilities
faker>=19.3.0  # For generating test data
factory-boy>=3.3.0  # Test fixtures
hypothesis>=6.82.0  # Property-based testing

# Performance testing
pytest-benchmark>=4.0.0
memory-profiler>=0.61.0
py-spy>=0.3.14  # Python profiler

# API testing
httpx>=0.24.1  # For async HTTP testing
requests-mock>=1.11.0

# Database testing
pytest-sqlite>=0.2.0
sqlalchemy-utils>=0.41.1

# Security testing
bandit>=1.7.5  # Security linter
safety>=2.3.5  # Check dependencies for vulnerabilities

# Mocking and fixtures
responses>=0.23.3  # Mock HTTP responses
freezegun>=1.2.2  # Mock datetime

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.4.1
pylint>=2.17.5

# Documentation
pytest-html>=3.2.0  # HTML test reports
allure-pytest>=2.13.2  # Allure test reports

# Coverage
coverage[toml]>=7.3.0
pytest-cov>=4.1.0

# Other utilities
psutil>=5.9.5  # System monitoring
numpy>=1.24.3  # For test data generation
opencv-python>=4.8.0  # For image testing
Pillow>=10.0.0  # Image manipulation